Remove old runtime BOM backend code using open files
