+++ deepfence_backend/api/vulnerability_api.py
@@ -9,7 +9,7 @@ from datetime import datetime
 import urllib.parse
 import magic
 from utils.helper import websocketio_channel_name_format, split_list_into_chunks, get_recent_scan_ids, \
-    is_network_attack_vector, get_topology_network_graph, get_image_cve_status
+    is_network_attack_vector, get_topology_network_graph, get_image_cve_status, set_vulnerability_status_for_packages
 from utils.response import set_response, format_response
 from utils.common import calculate_interval, format_es_resp_for_api
 from utils.custom_exception import InvalidUsage, InternalError, DFError
@@ -19,7 +19,7 @@ from utils.constants import CVE_INDEX, TIME_UNIT_MAPPING, USER_ROLES, ES_TERMS_A
     NODE_TYPE_HOST, CVE_SCAN_TYPES, NODE_TYPE_CONTAINER_IMAGE, REGISTRY_IMAGES_CACHE_KEY_PREFIX, \
     MAX_TOTAL_SEVERITY_SCORE, MAX_TOP_EXPLOITABLE_VULNERABILITIES, REGISTRY_TYPE_GCLOUD, TOPOLOGY_FILTERS_PREFIX, \
     NODE_TYPE_REGISTRY_IMAGE, DF_ID_TO_SCOPE_ID_REDIS_KEY_PREFIX, NODE_ACTION_CVE_SCAN_START, ES_MAX_CLAUSE, \
-    CVE_SCAN_LOGS_INDEX, SBOM_INDEX, OPEN_FILES_CACHE_KEY, SBOM_DEFAULT_FIELDS
+    CVE_SCAN_LOGS_INDEX, SBOM_INDEX, SBOM_DEFAULT_FIELDS
 from utils.decorators import user_permission, non_read_only_user
 from models.node_tags import NodeTags
 from models.container_image_registry import RegistryCredential
@@ -38,16 +38,6 @@ from utils.scope import fetch_topology_hosts, fetch_topology_container_images
 
 vulnerability_api = Blueprint("vulnerability_api", __name__)
 vulnerability_api_restful = Api(vulnerability_api)
-
-
-SEVERITY_RANK = {
-    "critical": 0,
-    "high": 1,
-    "medium": 2,
-    "low": 3,
-    "info": 4,
-    "never_scanned": 5
-}
 
 
 @vulnerability_api.route("/vulnerabilities/cve_severity_chart", methods=["GET", "POST"])
@@ -1643,86 +1633,6 @@ def list_process_open_files():
 
     set_vulnerability_status_for_packages(open_files_list)
     return set_response(data={"data": list(open_files_list.values())})
-
-
-def set_vulnerability_status_for_packages(open_files_list):
-    file_severity_map = {}
-    package_queries = []
-    for package_key, open_files in open_files_list.items():
-        package_subqueries = [{
-            "terms": {
-                "cve_caused_by_package.keyword": [open_files["package"]]
-            }
-        },
-        {
-            "terms": {
-                "cve_caused_by_package_path.keyword": [open_files["filepath"]]
-            }
-        }]
-        container_image_names = set()
-        host_names = []
-        for file_host_detail in open_files["hosts"]:
-            host_names.append(file_host_detail["host_name"])
-            container_image_names.add(file_host_detail["host_name"])
-            container_image_names.update({container_detail["image_name"] for container_detail in file_host_detail["containers"]})
-        package_subqueries.append({
-            "terms": {
-                "host_name.keyword": host_names
-            }
-        })
-        package_subqueries.append({
-            "terms": {
-                "cve_container_image.keyword": list(container_image_names)
-            }
-        })
-        package_queries.append({
-            "bool": {
-                "must": package_subqueries,
-                "_name": package_key
-            }
-        })
-    query = {
-        "query": {
-            "bool": {
-                "should": package_queries
-            }
-        },
-        "sort": [
-            {
-                "cve_caused_by_package.keyword": "asc"
-            },
-            {
-                "_script": {
-                    "type": "number",
-                    "script": {
-                        "lang": "painless",
-                        "source": "params.sortOrder.indexOf(doc['cve_severity.keyword'].value)",
-                        "params": {"sortOrder": ["info", "low", "medium", "high", "critical"]}
-                    },
-                    "order": "desc"
-                }
-            },
-            {
-                "cve_overall_score": "desc"
-            }
-        ]
-    }
-    cve_es_list = ESConn.search(CVE_INDEX, query, 0, 49999)
-    cve_hits = []
-    if cve_es_list.get('hits', {}).get('total', {}).get('value', 0) != 0:
-        cve_hits = cve_es_list['hits']['hits']
-    for cve_hit in cve_hits:
-        new_severity = cve_hit["_source"]["cve_severity"]
-        for package_key in cve_hit["matched_queries"]:
-            if file_severity_map.get(package_key, None):
-                existing_severity = file_severity_map[package_key]
-                if SEVERITY_RANK[existing_severity] > SEVERITY_RANK[new_severity]:
-                    file_severity_map[package_key] = new_severity
-            else:
-                file_severity_map[package_key] = new_severity
-    for package_key, open_files in open_files_list.items():
-        if file_severity_map.get(package_key):
-            open_files_list[package_key]["vulnerability_status"] = file_severity_map[package_key]
 
 
 @vulnerability_api.route("/vulnerability/top_exploits_bubble_chart", methods=["GET"])
