+++ deepfence_backend/utils/constants.py
@@ -168,8 +168,6 @@ PDF_REPORT_MAX_DOCS = 75000
 REGISTRY_IMAGES_CACHE_KEY_PREFIX = 'REGISTRY_IMAGES_LIST'
 REGISTRY_IMAGES_CACHE_EXPIRY_TIME = datetime.timedelta(days=2)
 
-OPEN_FILES_CACHE_KEY = 'OPEN_FILES_LIST'
-OPEN_FILES_CACHE_EXPIRY_TIME = datetime.timedelta(days=2)
 SECRET_SCANNER_FILE_PREFIX = '/tmp/Deepfence/SecretScanning'
 
 DURATION_IN_MINS = [
